+++ src/neo4j-arc/graph-visualization/GraphVisualizer/DefaultPanelContent/styled.tsx
