+++ src/neo4j-arc/graph-visualization/GraphVisualizer/NodeInspectorDrawer.tsx
