+++ src/neo4j-arc/graph-visualization/GraphVisualizer/NodeInspectorPanel.tsx
